feat(rmpc): withdraw command — agent-initiated gateway redemption with preflight and JSON result (#345)
* chore: initialize issue #312 worktree

* feat(rmpc): withdraw command — agent-initiated gateway redemption (#312)

Add `rmpc withdraw` subcommand that mirrors `rmpc deposit` with
withdraw-specific preflight: vault paused, share allowance(agent, gateway),
share balance. Builds and broadcasts gateway.withdraw(orderId, shares,
sourceVault, deadline, idempotencyKey), waits for receipt, decodes the
AgentWithdraw log, and emits stable JSON with assetsOut, assetRecipient,
txHash, blockNumber. Exits 2 on any hard preflight refusal.

- commands/withdraw.rs: full implementation with withdraw-specific preflight
- errors.rs: ErrVaultPaused, ErrWithdrawCapExceeded, ErrShareBalanceInsufficient,
  ErrShareAllowanceInsufficient, ErrAgentWithdrawLogMissing
- abi/RobotMoneyGateway.json: withdraw function + AgentWithdraw event ABI
- abi/MockVault.json: paused() view ABI for preflight checks
- testing/ethereum-testnet/e2e-rust/tests/withdraw.rs: suite-07 refusal scenarios
- suite-07 CI workflow: wire withdraw test binary with docker teardown
- Skill docs: document withdraw subcommand in SKILL.md and references/write.md

* fix(rmpc): skip USDC allowance/balance checks in withdraw preflight

The withdraw command reused Preflight::run() which includes checks 9-10
(USDC allowance and balance). These are deposit-specific; for withdraw the
relevant checks are vault share allowance/balance, handled separately by
withdraw_vault_preflight(). Add Preflight::run_gateway_only() that stops at
check 8, and call it from the withdraw command.

Also remove unused test imports (Config, SignerConfig, PathBuf) in the
withdraw tests module that caused rust-lint to fail.

* fix(fork-e2e): remove redundant as u64 casts in withdrawal tests

* fix(fork-e2e): remove unnecessary parentheses in withdrawal deadline assignments

---------

+++ testing/fork-e2e-rust/tests/withdrawal.rs
@@ -310,7 +310,7 @@ fn agent_withdrawal_happy_path() {
 
     // Agent deposits USDC via gateway.deposit() → shares land at agent.address.
     approve_usdc(&agent, usdc, gateway, deposit_amount);
-    let deadline = (now_secs + 300) as u64;
+    let deadline = now_secs + 300;
     let deposit_receipt = agent
         .send(
             gateway,
@@ -341,7 +341,7 @@ fn agent_withdrawal_happy_path() {
     // Agent calls gateway.withdraw(orderId, shares_held, vault, deadline, idempotencyKey).
     // The gateway pulls shares from the agent (transferFrom(agent → gateway)) and
     // redeems them, sending USDC to assetRecipient.
-    let withdraw_deadline = (now_secs + 300) as u64;
+    let withdraw_deadline = now_secs + 300;
     let withdraw_receipt = agent
         .send(
             gateway,
@@ -483,7 +483,7 @@ fn agent_withdrawal_redirect_blocked() {
 
     // Deposit so agent holds shares.
     approve_usdc(&agent, usdc, gateway, deposit_amount);
-    let deadline = (now_secs + 300) as u64;
+    let deadline = now_secs + 300;
     agent
         .send(
             gateway,
@@ -608,7 +608,7 @@ fn agent_withdrawal_window_cap() {
 
     // Deposit full_amount so agent holds full_amount shares.
     approve_usdc(&agent, usdc, gateway, full_amount);
-    let deadline = (now_secs + 300) as u64;
+    let deadline = now_secs + 300;
     agent
         .send(
             gateway,
@@ -629,7 +629,7 @@ fn agent_withdrawal_window_cap() {
     approve_vault_shares(&agent, vault, gateway, shares);
 
     // First withdrawal: half_amount shares — should succeed and exhaust the window cap.
-    let w1_deadline = (now_secs + 300) as u64;
+    let w1_deadline = now_secs + 300;
     let w1 = agent
         .send(
             gateway,
@@ -651,7 +651,7 @@ fn agent_withdrawal_window_cap() {
     );
 
     // Second withdrawal of even 1 share in same window — must revert.
-    let w2_deadline = (now_secs + 300) as u64;
+    let w2_deadline = now_secs + 300;
     let result = agent.send(
         gateway,
         &IGateway::withdrawCall {
